Exclude certs creation for ert3 tests
