+++ tests/ert_tests/conftest.py
@@ -2,6 +2,8 @@
 import logging
 import pytest
 
+from ert_shared.ensemble_evaluator.config import EvaluatorServerConfig
+
 
 @pytest.fixture(autouse=True)
 def log_check():
@@ -13,3 +15,21 @@ def log_check():
     assert (
         logging.WARNING == level_after
     ), f"Detected differences in log environment: Changed to {level_after}"
+
+
+@pytest.fixture(autouse=True)
+def no_cert_in_test(monkeypatch):
+    # Do not generate certificates during test, parts of it can be time
+    # consuming (e.g. 30 seconds)
+    # Specifically generating the RSA key <_openssl.RSA_generate_key_ex>
+    class MockESConfig(EvaluatorServerConfig):
+        def __init__(self, *args, **kwargs):
+            if "generate_cert" not in kwargs:
+                kwargs["generate_cert"] = False
+            super().__init__(*args, **kwargs)
+
+    for object_to_mock in [
+        "ert_shared.cli.main.EvaluatorServerConfig",
+        "ert.ert3.evaluator._evaluator.EvaluatorServerConfig",
+    ]:
+        monkeypatch.setattr(object_to_mock, MockESConfig)
